pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "1.0.0"
description = ""
authors = ["skip"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
coverage = "^5.4"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.1"
pytest = "^6.1.0"
coverage = "^5.4"
black = {version = "^20.8b1", allow-prereleases = true}
autopep8 = "^1.5.6"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"